Fix japanese filename handling for Nougat.
In Nougat, japanese folder name cause crash.
This is because after nougat, default charset for zip file is UTF8 and windows generated zip files is CP932.
There are no CP932 or CP942 charset in my envronment. So specify CP437.
I'm not sure whether this charset is available for all device, but previous version is crash anyway.